Add location and device info

+++ app/src/main/java/com/example/colorfight/data/userinfo/model/location/DeviceLocation.kt
@@ -1,4 +1,4 @@
 package com.example.colorfight.data.userinfo.model.location
 
-data class DeviceLocation(val lat: Float,
-						  val lon: Float)
+data class DeviceLocation(val lat: Double,
+						  val lon: Double)
