+++ app/src/main/java/com/example/colorfight/data/userinfo/model/location/DeviceLocationDTO.kt
@@ -1,4 +1,4 @@
 package com.example.colorfight.data.userinfo.model.location
 
-data class DeviceLocationDTO(val lat: Float,
-							 val lon: Float)
+data class DeviceLocationDTO(val lat: Double,
+							 val lon: Double)
